tools/eaasp-governance/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "eaasp-governance"
version = "0.1.0"
description = "EAASP L3 Governance Service — policy DSL compiler and governance API"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[tool.hatch.build.targets.wheel]
packages = ["src/eaasp_governance"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
